+++ src/debugger/manageddebugger.cpp
@@ -300,7 +300,7 @@ ManagedDebugger::ManagedDebugger()
       m_sharedEvalWaiter(new EvalWaiter),
       m_sharedEvalHelpers(new EvalHelpers(m_sharedEvalWaiter)),
       m_sharedEvalStackMachine(new EvalStackMachine),
-      m_sharedEvaluator(new Evaluator(m_sharedDebugInfo, m_sharedEvalHelpers, m_sharedEvalStackMachine)),
+      m_sharedEvaluator(new Evaluator(m_sharedDebugInfo, m_sharedEvalHelpers, m_sharedEvalStackMachine, m_sharedEvalWaiter)),
       m_sharedVariables(new Variables(m_sharedEvalHelpers, m_sharedEvaluator, m_sharedEvalStackMachine)),
       m_uniqueSteppers(new Steppers(m_sharedDebugInfo, m_sharedEvalHelpers)),
       m_sharedBreakpoints(new Breakpoints(m_sharedDebugInfo, m_sharedEvaluator, m_sharedVariables)),